warn about fullscreen in console
Corsica client makes an optimistic attempt to initiate fullscreen without a user action. In Firefox this creates an apparently uncatchable TypeError. This fix adds a note to the console logs before the attempt is made. It also adds a few lines to the README to explain different ways of launching in a kiosk mode for managed deployments.