WIP -- Dead End
Partial conversion. There are a lot of TypeErrors where the conversion
to an async message bus isn't exactly mapping to the promise/reduce
API that it replaced. Combined with the lack of tests, I suspect it
would be easier to toss all of this and rewrite and for version 3 using
the last 6 years as a living standard.

Commiting and pushing for posterity. Maybe and wrong and this can be
resumed someday.